+++ R/bin_time.R
@@ -123,7 +123,8 @@ bin_time <- function(occdf, min_ma = "min_ma", max_ma = "max_ma",
     if (is.data.frame(bins) == FALSE) {
       stop("`bins` should be a dataframe.")
     }
-    if (any(is.na(occdf[, max_ma])) || any(is.na(occdf[, min_ma]))) {
+    if (any(is.na(occdf[, max_ma, drop = TRUE])) ||
+        any(is.na(occdf[, min_ma, drop = TRUE]))) {
       stop(paste("NA values detected in", max_ma, "or", min_ma))
     }
 
@@ -148,13 +149,15 @@ bin_time <- function(occdf, min_ma = "min_ma", max_ma = "max_ma",
                   " column(s) do not exist in `bins`."))
     }
 
-    if (is.numeric(occdf[, max_ma]) &&
-        max(occdf[, max_ma]) > max(bins[, max_ma])) {
+    if (is.numeric(occdf[, max_ma, drop = TRUE]) &&
+        max(occdf[, max_ma, drop = TRUE]) >
+        max(bins[, max_ma, drop = TRUE])) {
       stop("Maximum age of occurrence data surpasses maximum age of bins.")
     }
 
-    if (is.numeric(occdf[, min_ma]) &&
-        min(occdf[, min_ma]) < min(bins[, min_ma])) {
+    if (is.numeric(occdf[, min_ma, drop = TRUE]) &&
+        min(occdf[, min_ma, drop = TRUE]) <
+        min(bins[, min_ma, drop = TRUE])) {
       stop("Minimum age of occurrence data is less than minimum age of bins.")
     }
 
@@ -171,8 +174,8 @@ bin_time <- function(occdf, min_ma = "min_ma", max_ma = "max_ma",
     # For each occurrence, find all the bins that it is present within, and
     # add as elements to that part of the list.
     for (i in seq_len(nrow(bins))) {
-      v <- which(occdf[, max_ma] > bins[i, min_ma] &
-                 occdf[, min_ma] < bins[i, max_ma])
+      v <- which(occdf[, max_ma, drop = TRUE] > bins[i, min_ma, drop = TRUE] &
+                 occdf[, min_ma, drop = TRUE] < bins[i, max_ma, drop = TRUE])
       for (j in v) {
         bin_list[[j]] <- append(bin_list[[j]], bins$bin[i])
       }
@@ -192,7 +195,8 @@ bin_time <- function(occdf, min_ma = "min_ma", max_ma = "max_ma",
     occdf$n_bins <- lengths(bin_list)
 
     # Generate midpoint ages of bins
-    bins$mid_ma <- (bins[, max_ma] + bins[, min_ma]) / 2
+    bins$mid_ma <- (bins[, max_ma, drop = TRUE] +
+                      bins[, min_ma, drop = TRUE]) / 2
 
     #=== Methods ===
 
@@ -202,7 +206,8 @@ bin_time <- function(occdf, min_ma = "min_ma", max_ma = "max_ma",
       # new column.
       rmcol <- FALSE
       if (("mid_ma" %in% colnames(occdf)) == FALSE) {
-        occdf$mid_ma <- (occdf[, max_ma] + occdf[, min_ma]) / 2
+        occdf$mid_ma <- (occdf[, max_ma, drop = TRUE] +
+                           occdf[, min_ma, drop = TRUE]) / 2
         rmcol <- TRUE
       }
       # Check if mid_ma equivalent to any bin boundaries
@@ -215,8 +220,8 @@ bin_time <- function(occdf, min_ma = "min_ma", max_ma = "max_ma",
 
       # Assign bin based on midpoint age of the age range
       for (i in seq_len(nrow(bins))) {
-        v <- which(occdf$mid_ma > bins[i, min_ma] &
-                   occdf$mid_ma < bins[i, max_ma])
+        v <- which(occdf$mid_ma > bins[i, min_ma, drop = TRUE] &
+                   occdf$mid_ma < bins[i, max_ma, drop = TRUE])
         occdf$bin_assignment[v] <- bins$bin[i]
         occdf$bin_midpoint[v] <- bins$mid_ma[i]
       }
@@ -260,8 +265,8 @@ bin_time <- function(occdf, min_ma = "min_ma", max_ma = "max_ma",
       # sample from it. Record that with each occurrence.
       for (i in seq_len(nrow(occdf))) {
         #generate occurrence sequence for sampling
-        occ_seq <- seq(from = occdf[i, min_ma],
-                       to = occdf[i, max_ma],
+        occ_seq <- seq(from = occdf[i, min_ma, drop = TRUE],
+                       to = occdf[i, max_ma, drop = TRUE],
                        by = 0.001)
         #generate x for input probability function
         x_prob <- seq(from = 0, to = 1, length.out = length(occ_seq))
@@ -289,8 +294,8 @@ bin_time <- function(occdf, min_ma = "min_ma", max_ma = "max_ma",
       for (i in seq_len(reps)) {
         occdf$point_estimates <- do.call(rbind, occ_list)[, i]
           for (j in seq_len(nrow(bins))){
-            vec <- which(occdf$point_estimates <= bins[j, max_ma] &
-                    occdf$point_estimates >= bins[j, min_ma])
+            vec <- which(occdf$point_estimates <= bins[j, max_ma, drop = TRUE] &
+                    occdf$point_estimates >= bins[j, min_ma, drop = TRUE])
             occdf$bin_assignment[vec] <- bins$bin[j]
           }
         occ_df_list[[i]] <- occdf
@@ -316,7 +321,7 @@ bin_time <- function(occdf, min_ma = "min_ma", max_ma = "max_ma",
         occdf$bin_midpoint[vec] <- bins$mid_ma[i]
       }
 
-      rownames(occdf) <- seq_len(nrow(occdf))
+      if (!inherits(occdf, "tbl")) rownames(occdf) <- seq_len(nrow(occdf))
 
       # Return the dataframe and end the function.
       return(occdf)
@@ -333,7 +338,8 @@ bin_time <- function(occdf, min_ma = "min_ma", max_ma = "max_ma",
         tmpbin <- bins[bins$bin %in% bin_list[[i]], ]
 
         # Generate sequence of length 10000 for percentage calculations
-        occ_seq <- seq(occdf[i, min_ma], occdf[i, max_ma], length.out = 10000)
+        occ_seq <- seq(occdf[i, min_ma, drop = TRUE],
+                       occdf[i, max_ma, drop = TRUE], length.out = 10000)
 
         # Calculate overlap across known bins
         percentage <- vector()