+++ R/tax_range_strat.R
@@ -119,7 +119,7 @@ tax_range_strat <- function(occdf, name = "genus", group = NULL, level = "bed",
     stop('`group` is not a named column in `occdf`')
   }
 
-  if (!is.numeric(occdf[, level])) {
+  if (!is.numeric(occdf[, level, drop = TRUE])) {
     stop("`level` must be of class numeric")
   }
 
@@ -134,16 +134,16 @@ tax_range_strat <- function(occdf, name = "genus", group = NULL, level = "bed",
     if (certainty %in% colnames(occdf) == FALSE) {
       stop("`certainty` is not a named column in `occdf`")
     }
-    if (any(is.na(occdf[, certainty]))) {
+    if (any(is.na(occdf[, certainty, drop = TRUE]))) {
       stop("The `certainty` column contains NA values")
     }
   }
 
-  if (any(is.na(occdf[, name]))) {
+  if (any(is.na(occdf[, name, drop = TRUE]))) {
     stop("The `name` column contains NA values")
   }
 
-  if (any(is.na(occdf[, level]))) {
+  if (any(is.na(occdf[, level, drop = TRUE]))) {
     stop("The `level` column contains NA values")
   }
 
@@ -152,7 +152,7 @@ tax_range_strat <- function(occdf, name = "genus", group = NULL, level = "bed",
   }
 
   #List and order unique taxa
-  unique_taxa <- unique(occdf[, name])
+  unique_taxa <- unique(occdf[, name, drop = TRUE])
   unique_taxa <- unique_taxa[order(unique_taxa)]
 
   #Create object to hold information
@@ -167,15 +167,17 @@ tax_range_strat <- function(occdf, name = "genus", group = NULL, level = "bed",
 
   #Populate nested list
   for (i in seq_along(unique_taxa)) {
-    occ_filter <- occdf[(occdf[, name] == unique_taxa[i]), ]
+
+    occ_filter <- occdf[(occdf[, name, drop = TRUE] == unique_taxa[i]), ]
     ranges[i, 3] <- min(occ_filter[level])
     ranges[i, 4] <- max(occ_filter[level])
     if (!is.null(group)) {
       ranges[i, 2] <- occ_filter[1, group]
     }
+    
     #If uncertainty is used, fill second set of columns for certain IDs
     if (!is.null(certainty)) {
-      occ_filter <- occ_filter[(occ_filter[, certainty] == 1), ]
+      occ_filter <- occ_filter[(occ_filter[, certainty, drop = TRUE] == 1), ]
       if (nrow(occ_filter) == 0) {
         occ_filter[1, ] <- NA
       }
@@ -206,8 +208,8 @@ tax_range_strat <- function(occdf, name = "genus", group = NULL, level = "bed",
 
   #Obtain uncertain occurrences
   if (!is.null(certainty)) {
-    certain <- occdf[(occdf[, certainty] != 0), ]
-    uncertain <- occdf[(occdf[, certainty] == 0), ]
+    certain <- occdf[(occdf[, certainty, drop = TRUE] != 0), ]
+    uncertain <- occdf[(occdf[, certainty, drop = TRUE] == 0), ]
   }
 
   #Create plot
@@ -252,12 +254,12 @@ tax_range_strat <- function(occdf, name = "genus", group = NULL, level = "bed",
              col = cols[1], lty = ltys[1], lwd = lwds[1])
   }
   if (is.null(certainty)) {
-    points(y = occdf[, level], x = occdf$ID, pch = pchs[1],
+    points(y = occdf[, level, drop = TRUE], x = occdf$ID, pch = pchs[1],
            col = cols[1], bg = bgs[1], cex = cexs[1])
   } else {
-    points(y = certain[, level], x = certain$ID, pch = pchs[1],
+    points(y = certain[, level, drop = TRUE], x = certain$ID, pch = pchs[1],
            col = cols[1], bg = bgs[1], cex = cexs[1])
-    points(y = uncertain[, level], x = uncertain$ID, pch = pchs[2],
+    points(y = uncertain[, level, drop = TRUE], x = uncertain$ID, pch = pchs[2],
            col = cols[2], bg = bgs[2], cex = cexs[2])
   }
   # plot y-axis