+++ NEWS.md
@@ -1,6 +1,11 @@
-# palaeoverse 1.3.1
+# Development version
 
+* Several updates to palaeorotate
+  * Mantle reference frame models removed
+  * TorsvikCocks2017 model added
+  * Added updated reconstruction files (1 myr resolution)
 * Fixed point plotting bug in tax_range_strat
+* Made all functions tibble-friendly (#118)
 
 # palaeoverse 1.3.0